requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
lxml==5.1.0
outcome==1.3.0.post0
pillow==10.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
